refactor(primarynav): move subcomponent within component file
