+++ src/components/PrimaryNav/PrimaryNav.stories.tsx
@@ -2,12 +2,11 @@ import type { StoryObj, Meta } from '@storybook/react';
 import React from 'react';
 
 import { PrimaryNav } from './PrimaryNav';
-import PrimaryNavItem from '../PrimaryNavItem';
 
 export default {
   title: 'Components/PrimaryNav',
   component: PrimaryNav,
-  subcomponents: { PrimaryNavItem },
+  subcomponents: { 'PrimaryNav.Item': PrimaryNav.Item },
   decorators: [
     (Story) => (
       <div style={{ background: '#000' }}>
