--- src/components/PrimaryNavItem/PrimaryNavItem.module.css
@@ -1,63 +1,0 @@
-@import '../../design-tokens/mixins.css';
-
-/*------------------------------------*\
-    # PRIMARY NAVIGATION ITEM
-\*------------------------------------*/
-
-/**
- * Primary nav link
- *
- * Individual link within each primary nav item.
- */
-.primary-nav__link {
-  display: flex;
-  color: var(--eds-theme-color-text-neutral-default-inverse);
-  text-decoration: none;
-  white-space: nowrap;
-  padding: var(--eds-size-1);
-  border-radius: var(--eds-border-radius-md);
-  transition: background var(--eds-anim-fade-quick) var(--eds-anim-ease);
-
-  &:hover,
-  &:focus-visible {
-    background: var(--eds-theme-color-button-primary-brand-background-hover);
-  }
-
-  &:active {
-    background: var(--eds-theme-color-button-primary-brandbackground-active);
-  }
-
-  &:focus-visible {
-    @mixin focusInverted;
-  }
-
-  @supports not selector(:focus-visible) {
-    &:hover,
-    &:focus {
-      background: var(--eds-theme-color-button-primary-brand-background-hover);
-    }
-
-    &:focus {
-      @mixin focusInverted;
-    }
-  }
-
-  /**
-  * Primary nav link within active primary nav item
-  */
-  .primary-nav__item.eds-is-active & {
-    background: var(--eds-theme-color-button-primary-brand-background-active);
-  }
-
-  @media screen and (prefers-reduced-motion) {
-    transition: none;
-  }
-}
-
-/**
- * Primary nav icon
- */
-.primary-nav__icon {
-  margin-right: var(--eds-size-2-and-half);
-  color: var(--eds-theme-color-text-neutral-default-inverse);
-}
