--- src/components/PrimaryNavItem/index.ts
@@ -1,1 +1,0 @@
-export { PrimaryNavItem as default } from './PrimaryNavItem';
