+++ src/index.ts
@@ -112,4 +112,3 @@ export type { BannerProps } from './components/Banner';
  * consuming apps.
  */
 // export { default as PrimaryNav } from './components/PrimaryNav';
-// export { default as PrimaryNavItem } from './components/PrimaryNavItem';
